Feature/improve theme (#10)
* ✨ Add language and theme selector

* 👽️ Add hook useProfileData to subscribe in profile changes

* ✨ Use SnackBar when baby info update

* 🌐 Fix key of locales

* 💄 Refactor theme management and update styles in components
