+++ src/locales/resources/en/translation_old.json
@@ -1,0 +1,9 @@
+{
+  "settingsPage": {
+    "babyInformation": {
+      "submit": {
+        "sucess": "Baby information updated successfully"
+      }
+    }
+  }
+}
