+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./calculateAge";
+export * from "./theme";
